UI_framework/project_YT/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
curlify==2.2.1
execnet==2.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
playwright==1.50.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pytest==8.3.4
pytest-xdist==3.6.1
requests==2.32.3
typing_extensions==4.12.2
tdata==2025.1
urllib3==2.3.0
var==1.9.2